# swagger/model_payment.go
/*
 * Payment API
 *
 * Demo code
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Payment struct {

	Type_ string `json:"type,omitempty"`

	Id string `json:"id,omitempty"`

	Version float32 `json:"version,omitempty"`

	OrganisationId string `json:"organisation_id,omitempty"`

	Attributes *PaymentAttributes `json:"attributes,omitempty"`
}